+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fixed build issues and tests from updated packages
 - Fixed a couple of bugs in stripe for dead code, and transaction history being broken
 - Fixed database resolver paths after new prisma update on bare metal
+- Added scripts for datetime conversation for all sqlite database migrating from prisma 6 -> prisma 7
 
 ## v1.6.5 - Small Fixes, Smoother Tracking, Hindi Support